+++ examples/celery_app.rs
@@ -1,10 +1,9 @@
 #![allow(non_upper_case_globals)]
 
+use anyhow::Result;
 use async_trait::async_trait;
-use celery::error::TaskError;
-use celery::task::TaskResult;
+use celery::prelude::*;
 use env_logger::Env;
-use exitfailure::ExitFailure;
 use structopt::StructOpt;
 use tokio::time::{self, Duration};
 
@@ -17,10 +16,14 @@ fn add(x: i32, y: i32) -> TaskResult<i32> {
 // Demonstrates a task that raises an error, and also how to customize task options.
 // In this case we override the default `max_retries`.
 #[celery::task(max_retries = 3)]
-fn buggy_task() -> TaskResult<()> {
-    Err(TaskError::UnexpectedError(
-        "This error is part of the example: it is used to showcase error handling".into(),
-    ))
+async fn buggy_task() -> TaskResult<()> {
+    let data = tokio::fs::read("this-file-doesn't-exist")
+        .await
+        .with_unexpected_err(|| {
+            "This error is part of the example, it is used to showcase error handling"
+        })?;
+    println!("Read {} bytes", data.len());
+    Ok(())
 }
 
 // Demonstrates a long running IO-bound task. By increasing the prefetch count, an arbitrary
@@ -47,11 +50,15 @@ fn bound_task(task: &Self) {
 )]
 enum CeleryOpt {
     Consume,
-    Produce,
+    Produce {
+        #[structopt(possible_values = &["add", "buggy_task", "bound_task", "long_running_task"])]
+        tasks: Vec<String>,
+    },
 }
 
 #[tokio::main]
-async fn main() -> Result<(), ExitFailure> {
+async fn main() -> Result<()> {
+    env_logger::from_env(Env::default().default_filter_or("info")).init();
     let opt = CeleryOpt::from_args();
     let my_app = celery::app!(
         // broker = AMQP { std::env::var("AMQP_ADDR").unwrap_or_else(|_| "amqp://127.0.0.1:5672/my_vhost".into()) },
@@ -74,23 +81,38 @@ async fn main() -> Result<(), ExitFailure> {
 
     match opt {
         CeleryOpt::Consume => {
+            my_app.display_pretty().await;
             my_app.consume_from(&["celery", "buggy-queue"]).await?;
         }
-        CeleryOpt::Produce => {
-            // Basic task sending.
-            my_app.send_task(add::new(1, 2)).await?;
-            my_app.send_task(bound_task::new()).await?;
+        CeleryOpt::Produce { tasks } => {
+            if tasks.is_empty() {
+                // Basic task sending.
+                my_app.send_task(add::new(1, 2)).await?;
+                my_app.send_task(bound_task::new()).await?;
 
-            // Sending a task with additional options like `countdown`.
-            my_app.send_task(add::new(1, 3).with_countdown(3)).await?;
+                // Sending a task with additional options like `countdown`.
+                my_app.send_task(add::new(1, 3).with_countdown(3)).await?;
 
-            // Send the buggy task that will fail and be retried a few times.
-            my_app.send_task(buggy_task::new()).await?;
+                // Send the buggy task that will fail and be retried a few times.
+                my_app.send_task(buggy_task::new()).await?;
 
-            // Send the long running task that will fail with a timeout error.
-            my_app
-                .send_task(long_running_task::new(Some(3)).with_timeout(2))
-                .await?;
+                // Send the long running task that will fail with a timeout error.
+                my_app
+                    .send_task(long_running_task::new(Some(3)).with_time_limit(2))
+                    .await?;
+            } else {
+                for task in tasks {
+                    match task.as_str() {
+                        "add" => my_app.send_task(add::new(1, 2)).await?,
+                        "bound_task" => my_app.send_task(bound_task::new()).await?,
+                        "buggy_task" => my_app.send_task(buggy_task::new()).await?,
+                        "long_running_task" => {
+                            my_app.send_task(long_running_task::new(Some(3))).await?
+                        }
+                        _ => panic!("unknown task"),
+                    };
+                }
+            }
         }
     };